Formatted code; written more tests

+++ lib/src/complex.dart
@@ -13,38 +13,38 @@ class Complex implements Comparable<Complex> {
   final double imaginary;
 
   /// Creates a complex number with the given real and imaginary parts
-  const Complex(double real, double imaginary) :
-    real = real,
-    imaginary = imaginary;
+  const Complex(double real, double imaginary)
+      : real = real,
+        imaginary = imaginary;
 
   /// Creates a complex number having only the real part, meaning that the
   /// complex part is set to 0.
-  const Complex.fromReal(double real) :
-    real = real,
-    imaginary = 0;
+  const Complex.fromReal(double real)
+      : real = real,
+        imaginary = 0;
 
   /// Creates a complex number having only the imaginary part, meaning that the
   /// real part is set to 0.
-  const Complex.fromImaginary(double imaginary) :
-    real = 0,
-    imaginary = imaginary;
+  const Complex.fromImaginary(double imaginary)
+      : real = 0,
+        imaginary = imaginary;
 
   /// Creates a complex number by using a [Fraction] as parameter for the real
   /// and complex part.
-  Complex.fromFraction(Fraction real, Fraction imaginary) :
-    real = real.toDouble(),
-    imaginary = imaginary.toDouble();
+  Complex.fromFraction(Fraction real, Fraction imaginary)
+      : real = real.toDouble(),
+        imaginary = imaginary.toDouble();
 
   /// Creates a complex number having only the real part, which is expressed as
   /// a [Fraction]. The imaginary part is set to 0.
-  Complex.fromRealFraction(Fraction real) :
-        real = real.toDouble(),
+  Complex.fromRealFraction(Fraction real)
+      : real = real.toDouble(),
         imaginary = 0;
 
   /// Creates a complex number having only the imaginary part, which is expressed
   /// as a [Fraction]. The real part is set to 0.
-  Complex.fromImaginaryFraction(Fraction imaginary) :
-        real = 0,
+  Complex.fromImaginaryFraction(Fraction imaginary)
+      : real = 0,
         imaginary = imaginary.toDouble();
 
   /// Creates a complex number from the given polar coordinates where [r] is the
@@ -52,24 +52,28 @@ class Complex implements Comparable<Complex> {
   ///
   /// By default the angle theta must be expressed in *radians* but setting
   /// `angleInRadians = false` allows the value to be in degrees.
-  Complex.fromPolar(double r, double theta, {bool angleInRadians = true}) :
-    real = r * math.cos(angleInRadians ? theta : degToRad(theta)),
-    imaginary = r * math.sin(angleInRadians ? theta : degToRad(theta));
+  Complex.fromPolar(double r, double theta, {bool angleInRadians = true})
+      : real = r * math.cos(angleInRadians ? theta : degToRad(theta)),
+        imaginary = r * math.sin(angleInRadians ? theta : degToRad(theta));
 
   /// Creates a complex number by using a [Fraction] as parameter for the real
   /// and complex part.
-  const Complex.zero() : real = 0, imaginary = 0;
+  const Complex.zero()
+      : real = 0,
+        imaginary = 0;
 
   /// Creates the imaginary unit _i_ which is simply _0 + 1i_
-  const Complex.i() : real = 0, imaginary = 1;
+  const Complex.i()
+      : real = 0,
+        imaginary = 1;
 
   @override
-  bool operator==(Object other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
-      other is Complex &&
-      runtimeType == other.runtimeType &&
-      real == other.real &&
-      imaginary == other.imaginary;
+        other is Complex &&
+            runtimeType == other.runtimeType &&
+            real == other.real &&
+            imaginary == other.imaginary;
   }
 
   @override
@@ -90,11 +94,9 @@ class Complex implements Comparable<Complex> {
   /// Prints the complex number with opening and closing parenthesis in case the
   /// complex number has both real and imaginary part.
   String toStringWithParenthesis() {
-    if (real == 0)
-      return "${_fixZero(imaginary)}i";
+    if (real == 0) return "${_fixZero(imaginary)}i";
 
-    if (imaginary == 0)
-      return _fixZero(real);
+    if (imaginary == 0) return _fixZero(real);
 
     return "(${toString()})";
   }
@@ -102,10 +104,10 @@ class Complex implements Comparable<Complex> {
   /// Returns an instance of [PolarComplex] which contains the radius _r_ and the
   /// angle _phi_ of the complex number.
   PolarComplex toPolarCoordinates() => PolarComplex._(
-    r: abs,
-    phiRadians: phase,
-    phiDegrees: radToDeg(phase),
-  );
+        r: abs,
+        phiRadians: phase,
+        phiDegrees: radToDeg(phase),
+      );
 
   /// Converts the complex number into a string. If [asFraction] is `true` then
   /// the real and the imaginary part are converted into fractions rather than
@@ -121,14 +123,11 @@ class Complex implements Comparable<Complex> {
       imaginaryPartAbs = "${imaginary.abs().toFraction()}i";
     }
 
-    if (real == 0)
-      return imaginaryPart;
+    if (real == 0) return imaginaryPart;
 
-    if (imaginary == 0)
-      return realPart;
+    if (imaginary == 0) return realPart;
 
-    if (imaginary < 0)
-      return "$realPart - $imaginaryPartAbs";
+    if (imaginary < 0) return "$realPart - $imaginaryPartAbs";
 
     return "$realPart + $imaginaryPart";
   }
@@ -138,8 +137,7 @@ class Complex implements Comparable<Complex> {
   ///
   /// For example, `_fixZero(5.0)` returns `"5"`.
   String _fixZero(double value) {
-    if (value % 1 == 0)
-      return "${value.truncate()}";
+    if (value % 1 == 0) return "${value.truncate()}";
 
     return "$value";
   }
@@ -154,15 +152,15 @@ class Complex implements Comparable<Complex> {
   int compareTo(Complex other) => 0;
 
   /// Calculates the sum between two complex numbers.
-  Complex operator+(Complex other) =>
+  Complex operator +(Complex other) =>
       Complex(real + other.real, imaginary + other.imaginary);
 
   /// Calculates the difference between two complex numbers.
-  Complex operator-(Complex other) =>
+  Complex operator -(Complex other) =>
       Complex(real - other.real, imaginary - other.imaginary);
 
   /// Calculates the product of two complex numbers.
-  Complex operator*(Complex other) {
+  Complex operator *(Complex other) {
     var r = real * other.real - imaginary * other.imaginary;
     var i = real * other.imaginary + imaginary * other.real;
 
@@ -170,7 +168,7 @@ class Complex implements Comparable<Complex> {
   }
 
   /// Calculates the division of two complex numbers.
-  Complex operator/(Complex other) => this * other.reciprocal;
+  Complex operator /(Complex other) => this * other.reciprocal;
 
   /// Puts the opposite sign to the imaginary part of the complex number.
   Complex get conjugate => Complex(real, -imaginary);
@@ -184,8 +182,7 @@ class Complex implements Comparable<Complex> {
   ///  - (x - yi) / (x<sup>2</sup> + y<sup>2</sup>)
   ///  - conj(a + bi) / modulo(a + bi)<sup>2</sup>
   Complex get reciprocal {
-    if (isZero)
-      throw ComplexException("The reciprocal of zero is undefined.");
+    if (isZero) throw ComplexException("The reciprocal of zero is undefined.");
 
     var scale = real * real + imaginary * imaginary;
     return Complex(real / scale, -imaginary / scale);
@@ -197,7 +194,7 @@ class Complex implements Comparable<Complex> {
   /// It is the modulus/magnitude/absolute value of the complex number.
   double get abs {
     if ((real != 0) || (imaginary != 0))
-      return math.sqrt(real*real + imaginary*imaginary);
+      return math.sqrt(real * real + imaginary * imaginary);
 
     return 0;
   }
@@ -208,13 +205,16 @@ class Complex implements Comparable<Complex> {
 
   /// Calculates the _base-e_ exponential of a complex number z where _e_ is the
   /// famous Euler constant
-  Complex get exp => Complex(math.exp(real) * math.cos(imaginary), math.exp(real) * math.sin(imaginary));
+  Complex get exp => Complex(math.exp(real) * math.cos(imaginary),
+      math.exp(real) * math.sin(imaginary));
 
   /// Computes the sine of the complex number
-  Complex get sin => Complex(math.sin(real) * _cosh(imaginary), math.cos(real) * _sinh(imaginary));
+  Complex get sin => Complex(
+      math.sin(real) * _cosh(imaginary), math.cos(real) * _sinh(imaginary));
 
   /// Computes the cosine of the complex number
-  Complex get cos => Complex(math.cos(real) * _cosh(imaginary), -math.sin(real) * _sinh(imaginary));
+  Complex get cos => Complex(
+      math.cos(real) * _cosh(imaginary), -math.sin(real) * _sinh(imaginary));
 
   /// Computes the tangent of the complex number
   Complex get tan => sin / cos;
@@ -259,7 +259,6 @@ class Complex implements Comparable<Complex> {
       a = 1;
       b = 0;
     } else {
-
       if (n == 1) {
         a = real;
         b = imaginary;
@@ -267,8 +266,7 @@ class Complex implements Comparable<Complex> {
         var length = abs;
         var angle = phase;
 
-        if (angle < 0)
-          angle += math.pi * 2;
+        if (angle < 0) angle += math.pi * 2;
 
         length = math.pow(length, 1.0 / n) as double;
         angle = angle / n;
@@ -276,7 +274,6 @@ class Complex implements Comparable<Complex> {
         a = length * math.cos(angle);
         b = length * math.sin(angle);
       }
-
     }
 
     if (neg) {
@@ -293,12 +290,10 @@ class Complex implements Comparable<Complex> {
 
   /// The hyperbolic sine (sinh(x))
   double _sinh(num x) => (math.exp(x) - math.exp(-x)) / 2;
-  
 }
 
 /// Object returned by a [Complex] which represents the number in polar coordinates
 class PolarComplex {
-
   /// The absolute value/modulus of the complex number
   final double r;
 
@@ -310,10 +305,5 @@ class PolarComplex {
 
   /// Takes the radius [r] and the angle phi, expressed in both radians
   /// ([phiRadians]) and degrees ([phiDegrees]), of a complex number
-  const PolarComplex._({
-    this.r,
-    this.phiRadians,
-    this.phiDegrees
-  });
-
+  const PolarComplex._({this.r, this.phiRadians, this.phiDegrees});
 }
